+++ spec/lib/generators/repository_generator_spec.rb
@@ -22,6 +22,14 @@ RSpec.describe ROM::Generators::RepositoryGenerator, type: :generator do
                 commands :create, update: :by_pk, delete: :by_pk
 
                 struct_namespace Dummy
+
+                def by_id(id)
+                  root.by_pk(id).one
+                end
+
+                def all
+                  root.to_a
+                end
               end
             CONTENT
           end
@@ -44,6 +52,14 @@ RSpec.describe ROM::Generators::RepositoryGenerator, type: :generator do
                 commands :create, update: :by_pk, delete: :by_pk
 
                 struct_namespace Dummy
+
+                def by_id(id)
+                  root.by_pk(id).one
+                end
+
+                def all
+                  root.to_a
+                end
               end
             CONTENT
           end